+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - [Update] Update src/utils/image-processor.js. Add handling for empty or wrong parameter.
 - [Update] Given empty string as a fallback for article title.
 - [Update] Decrease redis expire time from 1 hour to 5 mins.
-- [Add] Add print icon.
+- [Feature] Add print icon for friendly printing.
 
 ### 1.1.3 
 - [Update] Change link of the animal protection project